Visualizing Earnings Based On College Majors
We work on the dataset on the job outcomes of students who graduated from college between 2010 and 2012. The original data on job outcomes was released by American Community Survey, which conducts surveys and aggregates the data. FiveThirtyEight cleaned the dataset and released it on their Github repo.
Using Matplotlib and seaborn libraries we aim to find:
1. Do students in more popular majors make more money?
2. How many majors are predominantly male or female?
3. Which category of majors have the most students?